requirements.txt
crewai
crewai_tools
langchain-openai
langchain-tavily
python-dotenv
pinecone-client
streamlit